refactor: CityResponse 재사용
