+++ backend/src/test/java/hanglog/trip/service/CityServiceTest.java
@@ -5,8 +5,8 @@ import static hanglog.trip.fixture.CityFixture.PARIS;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
 
+import hanglog.city.dto.response.CityResponse;
 import hanglog.trip.domain.repository.CityRepository;
-import hanglog.trip.dto.response.CityResponse;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -38,6 +38,6 @@ class CityServiceTest {
 
         // then
         assertThat(actual).usingRecursiveComparison()
-                .isEqualTo(List.of(CityResponse.of(PARIS), CityResponse.of(LONDON)));
+                .isEqualTo(List.of(CityResponse.withCountry(PARIS), CityResponse.withCountry(LONDON)));
     }
 }
